Added very simple reconnect logic
Right now, upon disconnect, the client that disconnected pauses and attempts to reconnect.
The server sends a pause and waits 5 seconds for reconnection before delcaring the remaining player the winner. Right now there is not a mechanism to ensure that the clients are still in sync - need to do this!